+++ codegen/impl/src/es/event/mod.rs
@@ -26,6 +26,14 @@ pub fn derive(input: TokenStream) -> syn::Result<TokenStream> {
 /// Helper attributes of `#[derive(Event)]` macro placed on an enum variant.
 #[derive(Debug, Default, ParseAttrs)]
 pub struct VariantAttrs {
+    /// Indicator whether this enum variant should be used as
+    /// [`event::Initialized`] rather than [`event::Sourced`].
+    ///
+    /// [`event::Initialized`]: arcana_core::es::event::Initialized
+    /// [`event::Sourced`]: arcana_core::es::event::Sourced
+    #[parse(ident, alias = initial)]
+    pub init: Option<syn::Ident>,
+
     /// Indicator whether to ignore this enum variant for code generation.
     #[parse(ident, alias = skip)]
     pub ignore: Option<syn::Ident>,
@@ -49,11 +57,14 @@ pub struct Definition {
     pub generics: syn::Generics,
 
     /// Single-[`Field`] [`Variant`]s of this enum to consider in code
-    /// generation.
+    /// generation, along with the indicator whether this variant should use
+    /// [`event::Initialized`] rather than [`event::Sourced`].
     ///
+    /// [`event::Initialized`]: arcana_core::es::event::Initialized
+    /// [`event::Sourced`]: arcana_core::es::event::Sourced
     /// [`Field`]: syn::Field
     /// [`Variant`]: syn::Variant
-    pub variants: Vec<syn::Variant>,
+    pub variants: Vec<(syn::Variant, bool)>,
 
     /// Indicator whether this enum has any variants marked with
     /// `#[event(ignore)]` attribute.
@@ -103,12 +114,25 @@ impl Definition {
     /// # Errors
     ///
     /// - If [`VariantAttrs`] failed to parse.
+    /// - If [`VariantAttrs::init`] and [`VariantAttrs::ignore`] were specified
+    ///   simultaneously.
     /// - If [`syn::Variant`] doesn't have exactly one unnamed 1 [`syn::Field`]
     ///   and is not ignored.
     fn parse_variant(
         variant: &syn::Variant,
-    ) -> syn::Result<Option<syn::Variant>> {
+    ) -> syn::Result<Option<(syn::Variant, bool)>> {
         let attrs = VariantAttrs::parse_attrs("event", variant)?;
+
+        if let Some(init) = &attrs.init {
+            if attrs.ignore.is_some() {
+                return Err(syn::Error::new(
+                    init.span(),
+                    "`init` and `ignore`/`skip` arguments are mutually \
+                     exclusive",
+                ));
+            }
+        }
+
         if attrs.ignore.is_some() {
             return Ok(None);
         }
@@ -126,7 +150,7 @@ impl Definition {
             ));
         }
 
-        Ok(Some(variant.clone()))
+        Ok(Some((variant.clone(), attrs.init.is_some())))
     }
 
     /// Substitutes the given [`syn::Generics`] with trivial types.
@@ -150,7 +174,7 @@ impl Definition {
     pub fn variant_types(&self) -> impl DoubleEndedIterator<Item = &syn::Type> {
         self.variants
             .iter()
-            .filter_map(|var| var.fields.iter().next().map(|f| &f.ty))
+            .filter_map(|var| var.0.fields.iter().next().map(|f| &f.ty))
     }
 
     /// Generates code to derive [`Event`][0] trait, by simply matching over
@@ -163,7 +187,7 @@ impl Definition {
         let ty = &self.ident;
         let (impl_gens, ty_gens, where_clause) = self.generics.split_for_impl();
 
-        let var = self.variants.iter().map(|v| &v.ident).collect::<Vec<_>>();
+        let var = self.variants.iter().map(|v| &v.0.ident).collect::<Vec<_>>();
 
         let unreachable_arm = self.has_ignored_variants.then(|| {
             quote! { _ => unreachable!(), }
@@ -175,9 +199,7 @@ impl Definition {
                 fn name(&self) -> ::arcana::es::event::Name {
                     match self {
                         #(
-                            Self::#var(f) => ::arcana::es::Event::name(
-                                ::arcana::es::event::codegen::Borrow::borrow(f),
-                            ),
+                            Self::#var(f) => ::arcana::es::Event::name(f),
                         )*
                         #unreachable_arm
                     }
@@ -186,9 +208,7 @@ impl Definition {
                 fn version(&self) -> ::arcana::es::event::Version {
                     match self {
                         #(
-                            Self::#var(f) => ::arcana::es::Event::version(
-                                ::arcana::es::event::codegen::Borrow::borrow(f),
-                            ),
+                            Self::#var(f) => ::arcana::es::Event::version(f),
                         )*
                         #unreachable_arm
                     }
@@ -208,8 +228,14 @@ impl Definition {
         let (_, ty_gens, _) = self.generics.split_for_impl();
         let turbofish_gens = ty_gens.as_turbofish();
 
-        let var = self.variants.iter().map(|v| &v.ident);
-        let var_ty = self.variant_types();
+        let var_ty = self.variants.iter().map(|(v, is_initial)| {
+            let var_ty = v.fields.iter().next().map(|f| &f.ty);
+            if *is_initial {
+                quote! { ::arcana::es::event::Initial<#var_ty> }
+            } else {
+                quote! { #var_ty }
+            }
+        });
 
         let mut ext_gens = self.generics.clone();
         ext_gens.params.push(parse_quote! { __S });
@@ -218,6 +244,24 @@ impl Definition {
         });
         let (impl_gens, _, where_clause) = ext_gens.split_for_impl();
 
+        let arms = self.variants.iter().map(|(v, is_initial)| {
+            let var = &v.ident;
+            let var_ty = v.fields.iter().next().map(|f| &f.ty);
+
+            let event = if *is_initial {
+                quote! {
+                    <::arcana::es::event::Initial<#var_ty>
+                     as ::arcana::RefCast>::ref_cast(f)
+                }
+            } else {
+                quote! { f }
+            };
+            quote! {
+                #ty#turbofish_gens::#var(f) => {
+                    ::arcana::es::event::Sourced::apply(self, #event);
+                },
+            }
+        });
         let unreachable_arm = self.has_ignored_variants.then(|| {
             quote! { _ => unreachable!(), }
         });
@@ -229,11 +273,7 @@ impl Definition {
             {
                 fn apply(&mut self, event: &#ty#ty_gens) {
                     match event {
-                        #(
-                            #ty#turbofish_gens::#var(f) => {
-                                ::arcana::es::event::Sourced::apply(self, f);
-                            },
-                        )*
+                        #( #arms )*
                         #unreachable_arm
                     }
                 }
@@ -259,7 +299,12 @@ impl Definition {
         let ty = &self.ident;
         let (impl_gens, ty_gens, where_clause) = self.generics.split_for_impl();
 
-        let var_ty = self.variant_types().collect::<Vec<_>>();
+        let var_ty = self
+            .variants
+            .iter()
+            .flat_map(|v| &v.0.fields)
+            .map(|f| &f.ty)
+            .collect::<Vec<_>>();
 
         // TODO: Use `Self::__arcana_events()` inside impl instead of type
         //       params substitution, once rust-lang/rust#57775 is resolved:
@@ -292,9 +337,7 @@ impl Definition {
 
                     let mut i = 0;
                     #({
-                        let events = <
-                            <#var_ty as #glue::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <#var_ty>::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -567,6 +610,7 @@ mod spec {
     fn derives_enum_impl() {
         let input = parse_quote! {
             enum Event {
+                #[event(init)]
                 File(FileEvent),
                 Chat(ChatEvent),
             }
@@ -577,23 +621,15 @@ mod spec {
             impl ::arcana::es::Event for Event {
                 fn name(&self) -> ::arcana::es::event::Name {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::name(f),
+                        Self::Chat(f) => ::arcana::es::Event::name(f),
                     }
                 }
 
                 fn version(&self) -> ::arcana::es::event::Version {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::version(f),
+                        Self::Chat(f) => ::arcana::es::Event::version(f),
                     }
                 }
             }
@@ -601,13 +637,19 @@ mod spec {
             #[automatically_derived]
             impl<__S> ::arcana::es::event::Sourced<Event> for Option<__S>
             where
-                Self: ::arcana::es::event::Sourced<FileEvent> +
+                Self: ::arcana::es::event::Sourced<
+                          ::arcana::es::event::Initial<FileEvent>
+                      > +
                       ::arcana::es::event::Sourced<ChatEvent>
             {
                 fn apply(&mut self, event: &Event) {
                     match event {
                         Event::File(f) => {
-                            ::arcana::es::event::Sourced::apply(self, f);
+                            ::arcana::es::event::Sourced::apply(
+                                self,
+                                <::arcana::es::event::Initial<FileEvent>
+                                 as ::arcana::RefCast>::ref_cast(f)
+                            );
                         },
                         Event::Chat(f) => {
                             ::arcana::es::event::Sourced::apply(self, f);
@@ -785,10 +827,7 @@ mod spec {
 
                     let mut i = 0;
                     {
-                        let events = <
-                            <FileEvent
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <FileEvent>::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -797,10 +836,7 @@ mod spec {
                         }
                     }
                     {
-                        let events = <
-                            <ChatEvent
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <ChatEvent>::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -833,6 +869,7 @@ mod spec {
     fn derives_enum_with_generics_impl() {
         let input = parse_quote! {
             enum Event<'a, F, C> {
+                #[event(init)]
                 File(FileEvent<'a, F>),
                 Chat(ChatEvent<'a, C>),
             }
@@ -843,23 +880,15 @@ mod spec {
             impl<'a, F, C> ::arcana::es::Event for Event<'a, F, C> {
                 fn name(&self) -> ::arcana::es::event::Name {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::name(f),
+                        Self::Chat(f) => ::arcana::es::Event::name(f),
                     }
                 }
 
                 fn version(&self) -> ::arcana::es::event::Version {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::version(f),
+                        Self::Chat(f) => ::arcana::es::Event::version(f),
                     }
                 }
             }
@@ -868,13 +897,19 @@ mod spec {
             impl<'a, F, C, __S> ::arcana::es::event::Sourced<Event<'a, F, C> >
                 for Option<__S>
             where
-                Self: ::arcana::es::event::Sourced<FileEvent<'a, F> > +
+                Self: ::arcana::es::event::Sourced<
+                          ::arcana::es::event::Initial<FileEvent<'a, F> >
+                      > +
                       ::arcana::es::event::Sourced<ChatEvent<'a, C> >
             {
                 fn apply(&mut self, event: &Event<'a, F, C>) {
                     match event {
                         Event::<'a, F, C>::File(f) => {
-                            ::arcana::es::event::Sourced::apply(self, f);
+                            ::arcana::es::event::Sourced::apply(
+                                self,
+                                <::arcana::es::event::Initial<FileEvent<'a, F> >
+                                 as ::arcana::RefCast>::ref_cast(f)
+                            );
                         },
                         Event::<'a, F, C>::Chat(f) => {
                             ::arcana::es::event::Sourced::apply(self, f);
@@ -1066,10 +1101,7 @@ mod spec {
 
                     let mut i = 0;
                     {
-                        let events = <
-                            <FileEvent<'a, F>
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <FileEvent<'a, F> >::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -1078,10 +1110,7 @@ mod spec {
                         }
                     }
                     {
-                        let events = <
-                            <ChatEvent<'a, C>
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <ChatEvent<'a, C> >::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -1134,24 +1163,16 @@ mod spec {
             impl ::arcana::es::Event for Event {
                 fn name(&self) -> ::arcana::es::event::Name {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::name(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::name(f),
+                        Self::Chat(f) => ::arcana::es::Event::name(f),
                         _ => unreachable!(),
                     }
                 }
 
                 fn version(&self) -> ::arcana::es::event::Version {
                     match self {
-                        Self::File(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
-                        Self::Chat(f) => ::arcana::es::Event::version(
-                            ::arcana::es::event::codegen::Borrow::borrow(f),
-                        ),
+                        Self::File(f) => ::arcana::es::Event::version(f),
+                        Self::Chat(f) => ::arcana::es::Event::version(f),
                         _ => unreachable!(),
                     }
                 }
@@ -1346,10 +1367,7 @@ mod spec {
 
                     let mut i = 0;
                     {
-                        let events = <
-                            <FileEvent
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <FileEvent>::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -1358,10 +1376,7 @@ mod spec {
                         }
                     }
                     {
-                        let events = <
-                            <ChatEvent
-                             as ::arcana::es::event::codegen::Unpacked>::Type
-                        >::__arcana_events();
+                        let events = <ChatEvent>::__arcana_events();
                         let mut j = 0;
                         while j < events.len() {
                             res[i] = events[j];
@@ -1450,6 +1465,23 @@ mod spec {
         assert_eq!(
             err.to_string(),
             "enum must have at least one non-ignored variant",
+        );
+    }
+
+    #[test]
+    fn errors_on_both_init_and_ignored_variant() {
+        let input = parse_quote! {
+            enum Event {
+                #[event(init, ignore)]
+                Event1(Event1),
+            }
+        };
+
+        let err = super::derive(input).unwrap_err();
+
+        assert_eq!(
+            err.to_string(),
+            "`init` and `ignore`/`skip` arguments are mutually exclusive",
         );
     }
 }